Add dedicated integration tests for the `publish local` sub-command
