+++ modules/integration/src/test/scala/scala/cli/integration/PublishLocalTests212.scala
@@ -1,0 +1,4 @@
+package scala.cli.integration
+
+class PublishLocalTests212
+    extends PublishLocalTestDefinitions(scalaVersionOpt = Some(Constants.scala212))
