+++ modules/integration/src/test/scala/scala/cli/integration/PublishLocalTests213.scala
@@ -1,0 +1,4 @@
+package scala.cli.integration
+
+class PublishLocalTests213
+    extends PublishLocalTestDefinitions(scalaVersionOpt = Some(Constants.scala213))
