+++ modules/integration/src/test/scala/scala/cli/integration/PublishLocalTestsDefault.scala
@@ -1,0 +1,3 @@
+package scala.cli.integration
+
+class PublishLocalTestsDefault extends PublishLocalTestDefinitions(scalaVersionOpt = None)
