+++ modules/integration/src/test/scala/scala/cli/integration/PublishTestsDefault.scala
@@ -3,115 +3,6 @@ package scala.cli.integration
 import com.eed3si9n.expecty.Expecty.expect
 
 class PublishTestsDefault extends PublishTestDefinitions(scalaVersionOpt = None) {
-
-  private object PublishTestInputs {
-    def testOrg     = "test-local-org.sth"
-    def testName    = "my-proj"
-    def testVersion = "1.5.6"
-    def sbv         = "2.13"
-    def projFile(message: String): String =
-      s"""//> using publish.organization "$testOrg"
-         |//> using publish.name "$testName"
-         |//> using publish.version "$testVersion"
-         |
-         |//> using scala "$sbv"
-         |//> using dep "com.lihaoyi::os-lib:0.8.1"
-         |
-         |object Project {
-         |  def message = "$message"
-         |
-         |  def main(args: Array[String]): Unit =
-         |    println(message)
-         |}
-         |""".stripMargin
-    val inputs = TestInputs(
-      os.rel / "Project.scala" -> projFile("Hello")
-    )
-  }
-
-  test("publish local") {
-    val expectedFiles = {
-      val modName = s"${PublishTestInputs.testName}_2.13"
-      val base    = os.rel / PublishTestInputs.testOrg / modName / PublishTestInputs.testVersion
-      val baseFiles = Seq(
-        base / "jars" / s"$modName.jar",
-        base / "docs" / s"$modName-javadoc.jar",
-        base / "srcs" / s"$modName-sources.jar",
-        base / "poms" / s"$modName.pom",
-        base / "ivys" / "ivy.xml"
-      )
-      baseFiles
-        .flatMap { f =>
-          val md5  = f / os.up / s"${f.last}.md5"
-          val sha1 = f / os.up / s"${f.last}.sha1"
-          Seq(f, md5, sha1)
-        }
-        .toSet
-    }
-
-    PublishTestInputs.inputs.fromRoot { root =>
-      os.proc(
-        TestUtil.cli,
-        "--power",
-        "publish",
-        "local",
-        "Project.scala",
-        "--ivy2-home",
-        os.rel / "ivy2"
-      )
-        .call(cwd = root)
-      val ivy2Local = root / "ivy2" / "local"
-      val foundFiles = os.walk(ivy2Local)
-        .filter(os.isFile(_))
-        .map(_.relativeTo(ivy2Local))
-        .toSet
-      val missingFiles    = expectedFiles -- foundFiles
-      val unexpectedFiles = foundFiles -- expectedFiles
-      if (missingFiles.nonEmpty)
-        pprint.err.log(missingFiles)
-      if (unexpectedFiles.nonEmpty)
-        pprint.err.log(unexpectedFiles)
-      expect(missingFiles.isEmpty)
-      expect(unexpectedFiles.isEmpty)
-    }
-  }
-
-  test("publish local twice") {
-    PublishTestInputs.inputs.fromRoot { root =>
-      def publishLocal(): os.CommandResult =
-        os.proc(
-          TestUtil.cli,
-          "--power",
-          "publish",
-          "local",
-          "Project.scala",
-          "--ivy2-home",
-          os.rel / "ivy2",
-          "--working-dir",
-          os.rel / "work-dir"
-        )
-          .call(cwd = root)
-      def output(): String =
-        os.proc(
-          TestUtil.cs,
-          s"-J-Divy.home=${root / "ivy2"}",
-          "launch",
-          s"${PublishTestInputs.testOrg}:${PublishTestInputs.testName}_${PublishTestInputs.sbv}:${PublishTestInputs.testVersion}"
-        )
-          .call(cwd = root)
-          .out.trim()
-
-      publishLocal()
-      val output1 = output()
-      expect(output1 == "Hello")
-
-      os.write.over(root / "Project.scala", PublishTestInputs.projFile("olleH"))
-      publishLocal()
-      val output2 = output()
-      expect(output2 == "olleH")
-    }
-  }
-
   test("Pure Java") {
     val testOrg     = "test-org.foo"
     val testName    = "foo"
